+++ djoser/utils.py
@@ -1,4 +1,4 @@
-from django.contrib.auth import user_logged_in, user_logged_out
+from django.contrib.auth import user_logged_in, user_logged_out, login, logout
 from django.utils.encoding import force_bytes, force_text
 from django.utils.http import urlsafe_base64_encode, urlsafe_base64_decode
 
@@ -6,7 +6,7 @@ from djoser.conf import settings
 
 
 def encode_uid(pk):
-    return urlsafe_base64_encode(force_bytes(pk)).decode()
+    return force_text(urlsafe_base64_encode(force_bytes(pk)))
 
 
 def decode_uid(pk):
@@ -15,6 +15,8 @@ def decode_uid(pk):
 
 def login_user(request, user):
     token, _ = settings.TOKEN_MODEL.objects.get_or_create(user=user)
+    if settings.CREATE_SESSION_ON_LOGIN:
+        login(request, user)
     user_logged_in.send(sender=user.__class__, request=request, user=user)
     return token
 
@@ -25,10 +27,12 @@ def logout_user(request):
         user_logged_out.send(
             sender=request.user.__class__, request=request, user=request.user
         )
+    if settings.CREATE_SESSION_ON_LOGIN:
+        logout(request)
 
 
 class ActionViewMixin(object):
-    def post(self, request):
+    def post(self, request, **kwargs):
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         return self._action(serializer)