+++ djoser/views.py
@@ -2,7 +2,13 @@ from django.contrib.auth import get_user_model
 from django.contrib.auth.tokens import default_token_generator
 from django.urls.exceptions import NoReverseMatch
 from django.utils.timezone import now
-from rest_framework import generics, permissions, status, views, viewsets
+from rest_framework import (
+    generics,
+    permissions,
+    response,
+    status, views,
+    viewsets,
+)
 from rest_framework.decorators import list_route
 from rest_framework.response import Response
 from rest_framework.reverse import reverse
@@ -10,6 +16,7 @@ from rest_framework.reverse import reverse
 from djoser import utils, signals
 from djoser.compat import get_user_email, get_user_email_field_name
 from djoser.conf import settings
+from djoser.utils import ActionViewMixin
 
 User = get_user_model()
 
@@ -60,13 +67,7 @@ class RootView(views.APIView):
             return []
 
 
-class UserCreateView(generics.CreateAPIView):
-    """
-    Use this endpoint to register new user.
-    """
-    serializer_class = settings.SERIALIZERS.user_create
-    permission_classes = [permissions.AllowAny]
-
+class UserCreateMixin:
     def perform_create(self, serializer):
         user = serializer.save()
         signals.user_registered.send(
@@ -81,12 +82,53 @@ class UserCreateView(generics.CreateAPIView):
             settings.EMAIL.confirmation(self.request, context).send(to)
 
 
+class UserCreateView(UserCreateMixin, generics.CreateAPIView):
+    """
+    Use this endpoint to register new user.
+    """
+    serializer_class = settings.SERIALIZERS.user_create
+    permission_classes = settings.PERMISSIONS.user_create
+
+
+class ResendActivationView(ActionViewMixin, generics.GenericAPIView):
+    """
+    Use this endpoint to resend user activation email.
+    """
+    serializer_class = settings.SERIALIZERS.password_reset
+    permission_classes = [permissions.AllowAny]
+
+    _users = None
+
+    def _action(self, serializer):
+        if not settings.SEND_ACTIVATION_EMAIL:
+            return response.Response(status=status.HTTP_400_BAD_REQUEST)
+        for user in self.get_users(serializer.data['email']):
+            self.send_activation_email(user)
+        return response.Response(status=status.HTTP_204_NO_CONTENT)
+
+    def get_users(self, email):
+        if self._users is None:
+            email_field_name = get_user_email_field_name(User)
+            users = User._default_manager.filter(**{
+                email_field_name + '__iexact': email
+            })
+            self._users = [
+                u for u in users if not u.is_active and u.has_usable_password()
+            ]
+        return self._users
+
+    def send_activation_email(self, user):
+        context = {'user': user}
+        to = [get_user_email(user)]
+        settings.EMAIL.activation(self.request, context).send(to)
+
+
 class UserDeleteView(generics.CreateAPIView):
     """
     Use this endpoint to remove actually authenticated user
     """
     serializer_class = settings.SERIALIZERS.user_delete
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = settings.PERMISSIONS.user_delete
 
     def get_object(self):
         return self.request.user
@@ -107,7 +149,7 @@ class TokenCreateView(utils.ActionViewMixin, generics.GenericAPIView):
     Use this endpoint to obtain user authentication token.
     """
     serializer_class = settings.SERIALIZERS.token_create
-    permission_classes = [permissions.AllowAny]
+    permission_classes = settings.PERMISSIONS.token_create
 
     def _action(self, serializer):
         token = utils.login_user(self.request, serializer.user)
@@ -122,7 +164,7 @@ class TokenDestroyView(views.APIView):
     """
     Use this endpoint to logout user (remove user authentication token).
     """
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = settings.PERMISSIONS.token_destroy
 
     def post(self, request):
         utils.logout_user(request)
@@ -134,7 +176,7 @@ class PasswordResetView(utils.ActionViewMixin, generics.GenericAPIView):
     Use this endpoint to send email to user with password reset link.
     """
     serializer_class = settings.SERIALIZERS.password_reset
-    permission_classes = [permissions.AllowAny]
+    permission_classes = settings.PERMISSIONS.password_reset
 
     _users = None
 
@@ -164,7 +206,7 @@ class SetPasswordView(utils.ActionViewMixin, generics.GenericAPIView):
     """
     Use this endpoint to change user password.
     """
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = settings.PERMISSIONS.set_password
 
     def get_serializer_class(self):
         if settings.SET_PASSWORD_RETYPE:
@@ -185,7 +227,7 @@ class PasswordResetConfirmView(utils.ActionViewMixin, generics.GenericAPIView):
     """
     Use this endpoint to finish reset password process.
     """
-    permission_classes = [permissions.AllowAny]
+    permission_classes = settings.PERMISSIONS.password_reset_confirm
     token_generator = default_token_generator
 
     def get_serializer_class(self):
@@ -206,7 +248,7 @@ class ActivationView(utils.ActionViewMixin, generics.GenericAPIView):
     Use this endpoint to activate user account.
     """
     serializer_class = settings.SERIALIZERS.activation
-    permission_classes = [permissions.AllowAny]
+    permission_classes = settings.PERMISSIONS.activation
     token_generator = default_token_generator
 
     def _action(self, serializer):
@@ -230,7 +272,7 @@ class SetUsernameView(utils.ActionViewMixin, generics.GenericAPIView):
     """
     Use this endpoint to change user username.
     """
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = settings.PERMISSIONS.set_username
 
     def get_serializer_class(self):
         if settings.SET_USERNAME_RETYPE:
@@ -252,19 +294,9 @@ class SetUsernameView(utils.ActionViewMixin, generics.GenericAPIView):
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-class UserView(generics.RetrieveUpdateAPIView):
-    """
-    Use this endpoint to retrieve/update user.
-    """
-    model = User
-    serializer_class = settings.SERIALIZERS.user
-    permission_classes = [permissions.IsAuthenticated]
-
-    def get_object(self, *args, **kwargs):
-        return self.request.user
-
+class UserUpdateMixin:
     def perform_update(self, serializer):
-        super(UserView, self).perform_update(serializer)
+        super(UserUpdateMixin, self).perform_update(serializer)
         user = serializer.instance
         if settings.SEND_ACTIVATION_EMAIL and not user.is_active:
             context = {'user': user}
@@ -272,24 +304,37 @@ class UserView(generics.RetrieveUpdateAPIView):
             settings.EMAIL.activation(self.request, context).send(to)
 
 
-class UserViewSet(UserCreateView, viewsets.ModelViewSet):
+class UserView(UserUpdateMixin,
+               generics.RetrieveUpdateAPIView):
+    """
+    Use this endpoint to retrieve/update user.
+    """
+    queryset = User.objects.all()
+    serializer_class = settings.SERIALIZERS.user
+    permission_classes = settings.PERMISSIONS.user
+
+    def get_object(self, *args, **kwargs):
+        return self.request.user
+
+
+class UserViewSet(UserCreateMixin,
+                  UserUpdateMixin,
+                  viewsets.ModelViewSet):
     serializer_class = settings.SERIALIZERS.user
     queryset = User.objects.all()
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = settings.PERMISSIONS.user
     token_generator = default_token_generator
 
     def get_permissions(self):
-        if self.action in ['create', 'confirm']:
-            self.permission_classes = [permissions.AllowAny]
+        if self.action == 'create':
+            self.permission_classes = settings.PERMISSIONS.user_create
+        elif self.action == 'confirm':
+            self.permission_classes = settings.PERMISSIONS.activation
         elif self.action == 'list':
-            self.permission_classes = [permissions.IsAdminUser]
+            self.permission_classes = settings.PERMISSIONS.user_list
         return super(UserViewSet, self).get_permissions()
 
     def get_serializer_class(self):
-        if self.action == 'me':
-            # Use the current user serializer on 'me' endpoints
-            self.serializer_class = settings.SERIALIZERS.current_user
-
         if self.action == 'create':
             return settings.SERIALIZERS.user_create
 
@@ -308,18 +353,14 @@ class UserViewSet(UserCreateView, viewsets.ModelViewSet):
 
             return settings.SERIALIZERS.set_username
 
+        elif self.action == 'me':
+            # Use the current user serializer on 'me' endpoints
+            return settings.SERIALIZERS.current_user
+
         return self.serializer_class
 
     def get_instance(self):
         return self.request.user
-
-    def perform_update(self, serializer):
-        super(UserViewSet, self).perform_update(serializer)
-        user = serializer.instance
-        if settings.SEND_ACTIVATION_EMAIL and not user.is_active:
-            context = {'user': user}
-            to = [get_user_email(user)]
-            settings.EMAIL.activation(self.request, context).send(to)
 
     def perform_destroy(self, instance):
         utils.logout_user(self.request)