+++ testproject/testapp/tests/social/test_token_jwt.py
@@ -1,5 +1,6 @@
 from django.test import TestCase
-from rest_framework_jwt.serializers import VerifyJSONWebTokenSerializer
+
+from rest_framework_simplejwt.serializers import TokenVerifySerializer
 
 from djoser.social.token.jwt import TokenStrategy
 from ..common import create_user
@@ -11,6 +12,6 @@ class JWTStrategyTestCase(TestCase):
         res = TokenStrategy.obtain(user)
         self.assertEqual(res['user'], user)
 
-        data = {'token': res['token']}
-        serializer = VerifyJSONWebTokenSerializer(data=data)
+        data = {'token': res['access']}
+        serializer = TokenVerifySerializer(data=data)
         self.assertTrue(serializer.is_valid())