+++ CHANGELOG.rst
@@ -5,23 +5,57 @@ Change Log
 This document records all notable changes to djoser.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
---------------------
-`1.3.1` (2018-10-09)
---------------------
+---------------------
+`1.5.1`_ (2019-04-02)
+---------------------
+
+* Fixed a vulnerability of UserViewSet that allowed to create new accounts on wrong endpoint. (Thanks to @karazuba for reporting)
+* Past minor version since 1.2 will get a bugfix update and affected versions will be removed from PyPI to prevent affected versions from being installed.
+
+---------------------
+`1.5.0`_ (2019-03-05)
+---------------------
+
+* Added endpoint to resend activation email.
+* Added Polish and Georgian translations.
+* Fix missing **kwargs in ActionViewmixin.post() handler.
+* Fixed documentation.
+* Other small fixes.
+
+---------------------
+`1.4.0`_ (2019-01-09)
+---------------------
+
+* Introduced new framework for setting default permissions for certain views.
+  See :ref:`documentation<view-permission-settings>`.
+* Fix permissions regression introduced in 1.3.2.
+  Default permission for user-list view set to read-only, like in 1.3.2
+  (defaults to read-only like in 1.3.2).
+
+---------------------
+`1.3.2`_ (2018-12-05)
+---------------------
+
+* Fix vulnerability of user endpoints.
+* Fix issue  that appears on DRF 3.9+ on legacy `/me/` endpoint.
+
+---------------------
+`1.3.1`_ (2018-10-09)
+---------------------
 
 * Fix issue with circular import
 
---------------------
-`1.3.0` (2018-09-12)
---------------------
+---------------------
+`1.3.0`_ (2018-09-12)
+---------------------
 
 * Split user serializers (thanks to @joshua-s)
 * Add Django 2.1 to tox.ini
 * Update travis.yml
 
---------------------
-`1.2.0` (2018-07-23)
---------------------
+---------------------
+`1.2.0`_ (2018-07-23)
+---------------------
 
 * Refactor urls to use new RESTful ViewSets
 * Retain old urls for compatibility
@@ -277,3 +311,10 @@ few bugfixes / documentation updates. List of changes:
 .. _1.1.3: https://github.com/sunscrapers/djoser/compare/1.1.2...1.1.3
 .. _1.1.4: https://github.com/sunscrapers/djoser/compare/1.1.3...1.1.4
 .. _1.1.5: https://github.com/sunscrapers/djoser/compare/1.1.4...1.1.5
+.. _1.2.0: https://github.com/sunscrapers/djoser/compare/1.1.5...1.2.0
+.. _1.3.0: https://github.com/sunscrapers/djoser/compare/1.2.0...1.3.0
+.. _1.3.1: https://github.com/sunscrapers/djoser/compare/1.3.0...1.3.1
+.. _1.3.2: https://github.com/sunscrapers/djoser/compare/1.3.1...1.3.2
+.. _1.3.3: https://github.com/sunscrapers/djoser/compare/1.3.2...1.3.3
+.. _1.4.0: https://github.com/sunscrapers/djoser/compare/1.3.3...1.4.0
+.. _1.5.0: https://github.com/sunscrapers/djoser/compare/1.4.0...1.5.0